Refactor engine internals
